level2_knowledge_assistant/requirements.txt
requests
python-dotenv
chromadb
sentence-transformers
